Also updated.
2003-07-30  Jeffrey Stedfast  <fejj@ximian.com>

	* camel-mime-message.c (write_to_stream): Also updated.

	* camel-data-wrapper.c (write_to_stream): This should return ssize_t

	* camel-multipart-signed.c (write_to_stream): Updated.

	* camel-multipart.c (write_to_stream): Same.

	* camel-mime-part.c (write_to_stream): Here too.

svn path=/trunk/; revision=22029

+++ camel/camel-data-wrapper.c
@@ -41,7 +41,7 @@ static CamelObjectClass *parent_class = NULL;
 #define CDW_CLASS(so) CAMEL_DATA_WRAPPER_CLASS (CAMEL_OBJECT_GET_CLASS(so))
 
 static int construct_from_stream (CamelDataWrapper *, CamelStream *);
-static int write_to_stream (CamelDataWrapper *data_wrapper, CamelStream *stream);
+static ssize_t write_to_stream (CamelDataWrapper *data_wrapper, CamelStream *stream);
 static void set_mime_type (CamelDataWrapper *data_wrapper, const char *mime_type);
 static gchar *get_mime_type (CamelDataWrapper *data_wrapper);
 static CamelContentType *get_mime_type_field (CamelDataWrapper *data_wrapper);
@@ -111,10 +111,10 @@ camel_data_wrapper_get_type (void)
 	return type;
 }
 
-static int
+static ssize_t
 write_to_stream (CamelDataWrapper *data_wrapper, CamelStream *stream)
 {
-	int ret;
+	ssize_t ret;
 	
 	if (data_wrapper->stream == NULL) {
 		return -1;
